refactor: 외부 API 응답 DTO 폴더 이동
