+++ src/main/java/org/leisureup/global/response/external/base/Body.java
@@ -1,9 +1,11 @@
-package org.leisureup.global.response.external.tourapi;
+package org.leisureup.global.response.external.base;
 
+import com.fasterxml.jackson.annotation.*;
 import java.util.*;
 import lombok.extern.slf4j.*;
 
 @Slf4j
+@JsonIgnoreProperties(ignoreUnknown = true)
 public record Body<I>(
         int numOfRows,
         int pageNo,
