+++ src/main/java/org/leisureup/global/response/external/base/Items.java
@@ -1,9 +1,9 @@
-package org.leisureup.global.response.external.tourapi;
+package org.leisureup.global.response.external.base;
 
 import com.fasterxml.jackson.databind.annotation.*;
 import java.util.*;
 
-@JsonDeserialize(using = TourApiItemsDeserializer.class)
+@JsonDeserialize(using = DefaultItemsDeserializer.class)
 public record Items<I>(
         List<I> item
 ) {
